+++ packages/cli/src/commands/algolia.ts
@@ -5,7 +5,7 @@ import * as commander from 'commander';
 import path from 'path';
 
 import { AlgoliaApi, CeramicApi, TutorialMetadata, TutorialIndex } from '@builderdao/apis';
-import { ProposalStateE, getProposalState } from '@builderdao-sdk/dao-program';
+import { ProposalStateE, getProposalState } from '@builderdao/program-tutorial';
 import { BuilderDaoConfig } from 'src/services';
 import { getClient } from 'src/client';
 import { log as _log } from 'src/utils';