+++ packages/apis/CHANGELOG.md
@@ -1,9 +1,17 @@
 # @builderdao/apis
 
+## 0.0.3
+
+### Patch Changes
+
+- This version has been bumped due to inconsistent versioning
+- Updated dependencies
+  - @builderdao/program-tutorial@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes
 
 - BuilderDAO initial npm release
 - Updated dependencies
-  - @builderdao-sdk/dao-program@0.0.2
+  - @builderdao/program-tutorial@0.0.2