+++ packages/cli/CHANGELOG.md
@@ -1,12 +1,23 @@
 # @builderdao/cli
 
+## 0.0.3
+
+### Patch Changes
+
+- This version has been bumped due to inconsistent versioning
+- Updated dependencies
+  - @builderdao/program-tutorial@0.0.3
+  - @builderdao/apis@0.0.3
+  - @builderdao/data@0.0.3
+  - @builderdao/md-utils@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes
 
 - BuilderDAO initial npm release
 - Updated dependencies
-  - @builderdao-sdk/dao-program@0.0.2
+  - @builderdao/program-tutorial@0.0.2
   - @builderdao/apis@0.0.2
   - @builderdao/data@0.0.2
   - @builderdao/md-utils@0.0.2