+++ packages/md-formatter/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @builderdao/md-formatter
 
+## 0.0.3
+
+### Patch Changes
+
+- This version has been bumped due to inconsistent versioning
+- Updated dependencies
+  - @builderdao/cli@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes