+++ packages/md-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @builderdao/md-utils
 
+## 0.0.3
+
+### Patch Changes
+
+- This version has been bumped due to inconsistent versioning
+
 ## 0.0.2
 
 ### Patch Changes